Update with useReport hook
